+++ assets/js/client/debug_panel/debug_button.js
