+++ assets/js/client/debug_panel/debug_menu.js
